Half way through

+++ index.php
@@ -112,7 +112,7 @@
                     <td><?= $row['phone']; ?></td>
                     <td>
                        <a href="details.php?details=<?= $row['id']; ?>" class="badge badge-primary p-2">Details</a> 
-                       <a href="action.php?delete=<?= $row['id']; ?>" class="badge badge-danger p-2">Delete</a> 
+                       <a href="action.php?delete=<?= $row['id']; ?>" class="badge badge-danger p-2" onclick="return confirm('Are you sure you wanna delete this record?')">Delete</a> 
                        <a href="index.php?edit=<?= $row['id']; ?>" class="badge badge-success p-2">Edit</a> 
                     </td>
                 </tr>
